code cleanup

+++ src/tests/nps.rs
@@ -20,7 +20,7 @@ pub fn make_nps(positions: &str) -> NpsResult {
         searcher.search_to_time(&mut epd.position, 2000, false);
         let duration = start.elapsed();
         time += duration.as_millis();
-        nodes += searcher.get_nodes_searched() as u64;
+        nodes += searcher.get_nodes_searched();
     }
     NpsResult { time, nodes }
 }
